refactor(sessionhttp): split 4k-line workspace_handler.go by concern
workspace_handler.go was 4,097 lines. Relocate cohesive groups of
functions into sibling files within the same package (pure code
movement, no logic change):

- workspace_handler_groups.go  - group/move/trash/restore (16 funcs)
- workspace_handler_import.go  - workspace import & duplicate handling (18)
- workspace_handler_agents.go  - workspace agents, detail/analytics,
  layout (12)
- workspace_handler_sync.go    - sync/rescan/reconcile + disk hydration (15)

workspace_handler.go keeps the router, core CRUD, hydration helpers,
and all shared types/vars (24 funcs, 4097 -> 1227 lines). Total
top-level function count is unchanged (85).

Two pre-existing gosec findings moved into the new files (a workspace
import config read and the sync MkdirAll) get explicit #nosec
annotations with justifications, so the diff-based code-scanning gate
does not treat the relocated lines as new alerts. Behavior (including
the 0755 workspace-folder permission) is preserved.

Verified: go build ./..., go vet, gosec, and the full
./internal/sessionhttp/ test suite pass; gofmt clean.
